Move common namespace functions to common.go
