+++ instrumentor/controllers/deleteinstrumentationconfig/source_controller.go
@@ -24,10 +24,8 @@ import (
 	"github.com/odigos-io/odigos/k8sutils/pkg/consts"
 	k8sutils "github.com/odigos-io/odigos/k8sutils/pkg/utils"
 	"github.com/odigos-io/odigos/k8sutils/pkg/workload"
-
 	sourceutils "github.com/odigos-io/odigos/k8sutils/pkg/source"
-	appsv1 "k8s.io/api/apps/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
+
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -125,83 +123,5 @@ func (r *SourceReconciler) syncWorkload(ctx context.Context, source *v1alpha1.So
 		err = errors.Join(err, deleteWorkloadInstrumentationConfig(ctx, r.Client, obj))
 		err = errors.Join(err, removeReportedNameAnnotation(ctx, r.Client, obj))
 	}
-	return err
-}
-
-func syncNamespaceWorkloads(ctx context.Context, k8sClient client.Client, req ctrl.Request) error {
-	var err error
-	for _, kind := range []workload.WorkloadKind{
-		workload.WorkloadKindDaemonSet,
-		workload.WorkloadKindDeployment,
-		workload.WorkloadKindStatefulSet,
-	} {
-		err = errors.Join(err, listAndSyncWorkloadList(ctx, k8sClient, req, kind))
-	}
-	return err
-}
-
-func listAndSyncWorkloadList(ctx context.Context,
-	k8sClient client.Client,
-	req ctrl.Request,
-	kind workload.WorkloadKind) error {
-	logger := log.FromContext(ctx)
-	logger.V(2).Info("Uninstrumenting workloads for Namespace Source", "name", req.Name, "namespace", req.Namespace, "kind", kind)
-
-	workloads := workload.ClientListObjectFromWorkloadKind(kind)
-	err := k8sClient.List(ctx, workloads, client.InNamespace(req.Name))
-	if client.IgnoreNotFound(err) != nil {
-		return err
-	}
-
-	switch obj := workloads.(type) {
-	case *appsv1.DeploymentList:
-		for _, dep := range obj.Items {
-			err = syncGenericWorkloadListToNs(ctx, k8sClient, kind, client.ObjectKey{Namespace: dep.Namespace, Name: dep.Name})
-			if err != nil {
-				return err
-			}
-		}
-	case *appsv1.DaemonSetList:
-		for _, dep := range obj.Items {
-			err = syncGenericWorkloadListToNs(ctx, k8sClient, kind, client.ObjectKey{Namespace: dep.Namespace, Name: dep.Name})
-			if err != nil {
-				return err
-			}
-		}
-	case *appsv1.StatefulSetList:
-		for _, dep := range obj.Items {
-			err = syncGenericWorkloadListToNs(ctx, k8sClient, kind, client.ObjectKey{Namespace: dep.Namespace, Name: dep.Name})
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return err
-}
-
-func syncGenericWorkloadListToNs(ctx context.Context, c client.Client, kind workload.WorkloadKind, key client.ObjectKey) error {
-	// it is very important that we make the changes based on a fresh copy of the workload object
-	// if a list operation pulled in state and is now slowly iterating over it, we might be working with stale data
-	freshWorkloadCopy := workload.ClientObjectFromWorkloadKind(kind)
-	workloadGetErr := c.Get(ctx, key, freshWorkloadCopy)
-	if workloadGetErr != nil {
-		if apierrors.IsNotFound(workloadGetErr) {
-			// if the workload been deleted, we don't need to do anything
-			return nil
-		} else {
-			return workloadGetErr
-		}
-	}
-
-	instrumented, err := sourceutils.IsObjectInstrumentedBySource(ctx, c, freshWorkloadCopy)
-	if err != nil {
-		return err
-	}
-	if instrumented {
-		return nil
-	}
-
-	err = errors.Join(err, deleteWorkloadInstrumentationConfig(ctx, c, freshWorkloadCopy))
-	err = errors.Join(err, removeReportedNameAnnotation(ctx, c, freshWorkloadCopy))
 	return err
 }
